+++ vitepress-theme-default-plus.js
@@ -21,6 +21,7 @@ export {default as isActive} from './client/is-active.js';
 export {default as isDevRelease} from './utils/is-dev-release.js';
 export {default as isFauxInternal} from './utils/is-faux-internal.js';
 export {default as encodeTag} from './client/encode-tag.js';
+export {default as getBaseUrl} from './utils/get-base-url.js';
 export {default as useCollection} from './client/use-collection.js';
 export {default as useTags} from './client/use-tags.js';
 export {default as useTeam} from './client/use-team.js';