+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## {{ UNRELEASED_VERSION }} - [{{ UNRELEASED_DATE }}]({{ UNRELEASED_LINK }})
 
+* Added `RegEx` support for `internalDomains` matching
+* Added `is-faux-internal` helper util
+
 ## v1.1.0-beta.15 - [October 19, 2024](https://github.com/lando/vitepress-theme-default-plus/releases/tag/v1.1.0-beta.15)
 
 * Fixed bug with `<VPLLink>` not dynamically updating and showing strange behavior for `prev|next` links