+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `RegEx` support for `internalDomains` matching
 * Added `is-faux-internal` helper util
+* Standardized support envars to `VPL_` namespacing
 
 ## v1.1.0-beta.15 - [October 19, 2024](https://github.com/lando/vitepress-theme-default-plus/releases/tag/v1.1.0-beta.15)